Add 'swift_backup_url' back to cinder.conf
Removed as part of a cleanup to use FQDNs in place of IPs. However,
cinder swift driver does not support certificates other than those
acknowledged by curl or python requests. Thus we must connect
directly to swift without ssl. If we gather the url's from keystone
we are then trying to connect to the swift ssl port without using
ssl.